Add hit modifier for increasing headshot damage and potentially factoring armor placement

+++ NovaUnity/Assets/Src/Game/GamePlay/Attacking/AttackData.cs
@@ -4,12 +4,12 @@ public struct AttackData
 {
     public readonly string attackerUUID;
     public readonly int layer;
-    public readonly int potentialDamage;
+    public readonly float potentialDamage;
     public readonly DamageType damageType;
     public readonly Vector3 hitDirection;
     public readonly RaycastHit2D raycastHit;
     
-    public AttackData(string instigatorUUID, int attackLayer, DamageType type, int damage, Vector3 direction, RaycastHit2D potentialHit = default(RaycastHit2D))
+    public AttackData(string instigatorUUID, int attackLayer, DamageType type, float damage, Vector3 direction, RaycastHit2D potentialHit = default(RaycastHit2D))
     {
         attackerUUID = instigatorUUID;
         layer = attackLayer;
