+++ NovaUnity/Assets/Src/Game/GamePlay/Attacking/AttackResult.cs
@@ -2,11 +2,11 @@
 {
     public readonly AttackData attackData;
     public readonly IAttackTarget target;
-    public readonly int totalDamageDelt;
-    public readonly int targetHealthRemaining;
+    public readonly float totalDamageDelt;
+    public readonly float targetHealthRemaining;
     // public readonly bool hasKilledTarget;
 
-    public AttackResult(AttackData attackInfo, IAttackTarget attackTarget, int damage, int healthRemaining)
+    public AttackResult(AttackData attackInfo, IAttackTarget attackTarget, float damage, float healthRemaining)
     {
         attackData = attackInfo;
         target = attackTarget;
